+++ src/FilamentWidgetTabsServiceProvider.php
@@ -2,6 +2,8 @@
 
 namespace Shkubu\FilamentWidgetTabs;
 
+use Filament\Support\Assets\Css;
+use Filament\Support\Facades\FilamentAsset;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 
@@ -11,5 +13,12 @@ class FilamentWidgetTabsServiceProvider extends PackageServiceProvider
     {
         $package->name('filament-widget-tabs')
             ->hasViews();
+    }
+
+    public function packageBooted(): void
+    {
+        FilamentAsset::register([
+            Css::make('filament-widget-tabs', __DIR__ . '/../resources/dist/filament-widget-tabs.css'),
+        ], 'shkubu18/filament-widget-tabs');
     }
 }